% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ainv.R
\name{Ainv}
\alias{Ainv}
\title{Inverse of \eqn{A}}
\usage{
Ainv(x, kmax = 2000)
}
\arguments{
\item{x}{numeric value in \eqn{[0,1]}.}

\item{kmax}{positive value. Default value is 2000.}
}
\description{
\code{Ainv} returns the inverse of A, where A is the ratio of the first
  and zeroth order Bessel functions of the first kind with finite series
  approximation. If the inverse value \eqn{k} of \eqn{A(x)} is larger than
  \code{kmax}, then it returns \code{kmax}.
}
\examples{
\dontrun{
Ainv(0.7, kmax = 3000)
}
}
\references{
Mardia, K. V. and Jupp, P. E. (2000), Directional Statistics,
  New York: Wiley.
}
\seealso{
\code{\link[circular]{A1inv}}
}
