% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centers.torus.R
\name{centers.torus}
\alias{centers.torus}
\title{Centers from explicit kmeans clustering on torus}
\usage{
centers.torus(data, k = 10)
}
\arguments{
\item{data}{n x 2 matrix of toroidal data on \eqn{[0, 2\pi)^2}}

\item{k}{a scalar which determines the number of centers}
}
\value{
a matrix containing toroidal angles which indicates
  k centers.
}
\description{
\code{centers.torus} returns a matrix, containing toroidal angles
  which are given by explicit kmeans algorithm on torus.
}
\examples{
\dontrun{
## mean vectors

Mu1 <- c(3, 0)
Mu2 <- c(2, 2)
Mu3 <- c(1, 4)

## covariance matrices

Sigma1 <- matrix(c(0.1, 0.05, 0.05, 0.2), 2, 2)
Sigma2 <- matrix(c(0.1, 0, 0, 0.01), 2, 2)
Sigma3 <- matrix(c(0.01, 0, 0, 0.1), 2, 2)

## 2-dimensional multivariate normal data wrapped with toroidal space
require(MASS)
data <- rbind(mvrnorm(n=70, Mu1, Sigma1),
              mvrnorm(n=50, Mu2, Sigma2),
              mvrnorm(n=50, Mu3, Sigma3))
data <- on.torus(data)

centers.torus(data, k = 10)
}
}
\references{
S. Jung, K. Park, and B. Kim (2020),
  "Clustering on the torus by conformal prediction"
}
