% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.assign.number.R
\name{cluster.assign.number}
\alias{cluster.assign.number}
\title{Number of clusters for each ellipsoid numbers}
\usage{
cluster.assign.number(
  data,
  Jmin = 3,
  Jmax = 35,
  level = 0.1,
  split.id = NULL,
  method = c("homogeneous-circular", "heterogeneous-circular", "ellipsoids", "general"),
  init = c("kmeans", "hierarchical"),
  additional.condition = TRUE,
  THRESHOLD = 1e-10,
  maxiter = 200,
  verbose = FALSE
)
}
\arguments{
\item{data}{n x d matrix of toroidal data on \eqn{[0, 2\pi)^d}
or \eqn{[-\pi, \pi)^d}}

\item{Jmin}{minimum number of ellipsoids}

\item{Jmax}{maximum number of ellipsoids}

\item{level}{a scalar between \eqn{[0,1]}. Default value
is 0.1.}

\item{split.id}{a n-dimensinal vector consisting of values 1 (estimation)
and 2(evaluation)}

\item{method}{character which must be "homogeneous-circular",
"heterogeneous-circular", or "general".
 If "homogeneous-circular", the radii of k-spheres are identical.
 If "heterogeneous-cricular", the radii of k-spheres may be different.
 If "ellipsoids", cluster with k-ellipsoids without optimized parameters.
 If, "general", clustering with k-ellipses. The parameters to construct
 the ellipses are optimized with generalized Lloyd algorithm, which is
 modified for toroidal space. To see the detail, see the references.}

\item{init}{determine the initial parameter of "kmeans" method,
for option "general". Must be "kmeans" or "hierarchical".
If "kmeans", the initial parameters are obtained with extrinsic kmeans
method.
If "hierarchical", the initial parameters are obtained with hierarchical
clustering method.
Default is "kmeans".}

\item{additional.condition}{boolean index.
If \code{TRUE}, a singular matrix will be altered to the scaled identity.}

\item{THRESHOLD}{number for difference between updating and
updated parameters. Default is 1e-10.}

\item{maxiter}{the maximal number of iteration. Default is 200.}

\item{verbose}{boolean index, which indicates whether display
additional details as to what the algorithm is doing or
how many loops are done. Moreover, if \code{additional.condition} is
\code{TRUE}, the warning message will be reported.
Default is \code{FALSE}.}
}
\value{
an \code{output} object, containing a plot based on ggplot2,
  a \code{data.frame} for the number of clusters and icp.torus objects
  generated by given options, varied from \code{Jmin} to \code{Jmax}.
}
\description{
\code{cluster.assign.number} generates a table and a plot for the number of
  eliipsoids vs the number of clusters.
}
\examples{
\dontrun{
## mean vectors

Mu1 <- c(3, 0)
Mu2 <- c(2, 2)
Mu3 <- c(1, 4)

## covariance matrices

Sigma1 <- matrix(c(0.1, 0.05, 0.05, 0.2), 2, 2)
Sigma2 <- matrix(c(0.1, 0, 0, 0.01), 2, 2)
Sigma3 <- matrix(c(0.01, 0, 0, 0.1), 2, 2)

## 2-dimensional multivariate normal data wrapped with toroidal space
require(MASS)
data <- rbind(mvrnorm(n=70, Mu1, Sigma1),
              mvrnorm(n=50, Mu2, Sigma2),
              mvrnorm(n=50, Mu3, Sigma3))
data <- on.torus(data)

cluster.assign.number(data)
}

}
\references{
S. Jung, K. Park, and B. Kim (2020),
  "Clustering on the torus by conformal prediction", and
  Jaehyeok Shin, Alessandro Rinaldo and Larry Wasserman (2019),
  "Predictive Clustering"
}
\seealso{
\code{\link[ClusTorus]{kmeans.kspheres}}, \code{\link[ClusTorus]{icp.torus.score}}
}
