% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.assign.torus.R
\name{cluster.assign.torus}
\alias{cluster.assign.torus}
\title{Clustering by connected components of ellipsoids}
\usage{
cluster.assign.torus(data, icp.object, level = 0.1)
}
\arguments{
\item{data}{n x d matrix of toroidal data on \eqn{[0, 2\pi)^d}.}

\item{icp.object}{an object must be an \code{icp.torus} object, which contains
all values to compute the conformity score constructed with \code{icp.torus.score},
or a \code{hyperparam.torus} object which is generated by \code{hyperparam.torus}.}

\item{level}{a scalar in \eqn{[0,1]}. Default value
is 0.1. If argument \code{icp.object} is a \code{hyperparam.torus} object, then it is
automatically chosen by the optimal level.}
}
\value{
clustering assignment for data, given \code{icp.torus} objects
\describe{
  \item{\code{cluster.id.by.ehat}}{cluster assignment results which are based on log-approximation of max-mixture.}
  \item{\code{cluster.id.by.partialsum}}{cluster assignment results which are based on the mixture model.}
  \item{\code{cluster.id.outlier}}{cluster assignment results which regards data not included in conformal prediction set
  as outliers.}
  \item{\code{cluster.id.by.Mah.dist}}{cluster assignment results which are based on Mahalanobis distance.}
  \item{\code{level}}{used level which determines the size of clusters(conformal prediction set).}
  \item{\code{data}}{input data which are assigned to each cluster.}
  \item{\code{icp.torus}}{\code{icp.torus} object which is used for cluster assignment.}
}
}
\description{
\code{cluster.assign.torus} returns clustering assignment for data
  given \code{icp.torus} objects, which can be constructed with
  \code{icp.torus.score}.
}
\examples{
data <- toydata1[, 1:2]
icp.torus <- icp.torus.score(data, method = "kmeans",
                             kmeansfitmethod = "general",
                             J = 4, concentration = 25)
level <- 0.1

cluster.assign.torus(data, icp.torus, level = level)
}
\references{
S. Jung, K. Park, and B. Kim (2021),
  "Clustering on the torus by conformal prediction",

  I. Gilitschenski and U. D. Hanebeck,
  "A robust computational test for overlap of
  two arbitrary-dimensional ellipsoids in fault-detection of Kalman filters"
}
\seealso{
\code{\link[ClusTorus]{icp.torus.score}}, \code{\link[ClusTorus]{hyperparam.torus}}
}
