% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusTorus-package.R
\docType{package}
\name{ClusTorus-package}
\alias{ClusTorus}
\alias{ClusTorus-package}
\title{ClusTorus: Clustering on the Torus by Conformal Prediction}
\description{
Provides various tools of for clustering multivariate angular 
  data on the torus. The package provides angular 
  adaptations of usual clustering methods such as the k-means 
  clustering, pairwise angular distances, which can be used as an 
  input for distance-based clustering algorithms, and implements
  clustering based on the conformal prediction framework. Options 
  for the conformal scores include scores based on a kernel density 
  estimate, multivariate von Mises mixtures, and naive k-means clusters. 
  Moreover, the package provides some basic data handling tools for 
  angular data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sungkyujung/ClusTorus}
  \item Report bugs at \url{https://github.com/sungkyujung/ClusTorus/issues}
}

}
\author{
\strong{Maintainer}: Seungki Hong \email{skgaboja@snu.ac.kr}

Authors:
\itemize{
  \item Sungkyu Jung \email{sungkyu@snu.ac.kr} [copyright holder]
}

Other contributors:
\itemize{
  \item Kiho Park [contributor]
  \item Byungwon Kim [contributor]
}

}
\keyword{internal}
