% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.comp.ellipse.R
\name{conn.comp.ellipse}
\alias{conn.comp.ellipse}
\title{Connected components of given ellipses}
\usage{
conn.comp.ellipse(ellipse.param, t)
}
\arguments{
\item{ellipse.param}{list which is consisting of mean of each angular
coordinate, inverse of each covariance matrix, and constant term.}

\item{t}{a numeric value which determines the size of ellipses.}
}
\value{
a \code{list} which contains two components:

  \code{componentid} indicates which connected component
  the ellipse belongs to.
  \code{ncluster} indicates the number of clusters, i.e.,
  the number of connected components.
}
\description{
\code{conn.comp.ellipse} investigates how many connected
  components exist and which ellipses are included in
  each connected component.
}
\examples{
\dontrun{

parammat <- matrix(c(0.4, 0.3, 0.3,
                     20, 25, 25,
                     30, 25, 20,
                     1, 2, 3,
                     1, 2, 3,
                     0, 2, 4), nrow = 6, byrow =TRUE)

ellipse.param <- norm.appr.param(parammat)

t <- 0.5

conn.comp.ellipse(ellipse.param, t)
}
}
\references{
'S. Jung, K. Park, and B. Kim (2020),
  "Clustering on the torus by conformal prediction"
}
