% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehat.eval.R
\name{ehat.eval}
\alias{ehat.eval}
\title{Normal approximation to the log - weighted bivariate von Mises sine density}
\usage{
ehat.eval(X, ellipse.param)
}
\arguments{
\item{X}{n x 2 toroidal data on \eqn{[-\pi, \pi)^2}}

\item{ellipse.param}{list which is consisting of mean of each angular
coordinate, inverse of each covariance matrix, and constant term}
}
\value{
nrow(X) times ncol(parammat) (n x J) matrix
}
\description{
\code{ehat.eval} evaluates the approximation of
  log - weighted bivariate von Mises for each given data and f
  or each given parameters.
}
