% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMsinvMmix.init.R
\name{EMsinvMmix.init}
\alias{EMsinvMmix.init}
\title{Initalize for EMsinvMmix}
\usage{
EMsinvMmix.init(data, J = 4, hc = hclust(ang.pdist(data)))
}
\arguments{
\item{data}{n x 2 matrix of toroidal data on \eqn{[0, 2\pi)^2}}

\item{J}{number of components of mixture density}

\item{hc}{use \code{\link[stats]{hclust}} to construct
hierarchical clustering. You may tune the parameters of
\code{\link[stats]{hclust}} to change the method for
initalizing the parameters.}
}
\value{
returns \code{parammat} which will become the input of
  \code{\link[ClusTorus]{EMsinvMmix}}, the detail is below:
  6 x J parameter data with the following components:

  \code{parammat[1, ]} : the weights for each von Mises sine density

  \code{parammat[n + 1, ]} : \eqn{\kappa_n} for each von Mises
  sine density for n = 1, 2, 3

  \code{parammat[m + 4, ]} : \eqn{\mu_m} for each von Mises
   sine density for m = 1, 2
}
\description{
\code{EMsinvMmix.init} returns the inital parameter for the optimizing
  algorithm, \code{EMsinvMmix}.
}
\details{
This function uses complete-linkage hierarchical clustering
  method to set up the initial clusters, where the dissimilarity
  matrix may be constructed with \code{\link[ClusTorus]{ang.pdist}},
  which is the distance matrix for angular data. Then, use the
  membership to set mu, kappa by group-wise MLEs.
}
\examples{
\dontrun{
## mean vectors

Mu1 <- c(3, 0)
Mu2 <- c(2, 2)
Mu3 <- c(1, 4)

## covariance matrices

Sigma1 <- matrix(c(0.1, 0.05, 0.05, 0.2), 2, 2)
Sigma2 <- matrix(c(0.1, 0, 0, 0.01), 2, 2)
Sigma3 <- matrix(c(0.01, 0, 0, 0.1), 2, 2)

## 2-dimensional multivariate normal data wrapped with toroidal space
require(MASS)
data <- rbind(mvrnorm(n=70, Mu1, Sigma1),
              mvrnorm(n=50, Mu2, Sigma2),
              mvrnorm(n=50, Mu3, Sigma3))
data <- on.torus(data)

EMsinvMmix.init(data, J = 3)
}
}
\references{
'S. Jung, K. Park, and B. Kim (2020),
  "Clustering on the torus by conformal prediction"
}
\seealso{
See the detail for hierarchical clustering:
  \code{\link[stats]{hclust}}, \code{\link[ClusTorus]{ang.pdist}}.
   Also, see the detail of \code{\link[ClusTorus]{EMsinvMmix}}.
}
