% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.torus.R
\name{grid.torus}
\alias{grid.torus}
\title{Grid on torus}
\usage{
grid.torus(grid.size = 100)
}
\arguments{
\item{grid.size}{number of grid for each axis. Default value is 100.}
}
\value{
\code{grid.torus} returns (grid.size) x (grid.size) numeric matrix
  which indicates the grid points on torus
}
\description{
\code{grid.torus()} returns an equally-spaced grid on torus.
}
\examples{
\dontrun{
grid.torus()
}
}
