% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icp.torus.R
\name{icp.torus.score}
\alias{icp.torus.score}
\title{Conformity score for inductive prediction sets}
\usage{
icp.torus.score(
  data,
  split.id = NULL,
  method = c("all", "kde", "mixture"),
  mixturefitmethod = c("circular", "axis-aligned", "general", "Bayesian"),
  param = list(J = 4, concentration = 25)
)
}
\arguments{
\item{data}{n x 2 matrix of toroidal data on \eqn{[0, 2\pi)^2}}

\item{split.id}{a n-dimensinal vector consisting of values 1 (estimation)
and 2(evaluation)}

\item{method}{a string one of "all", "kde" and "mixture" which determines the
phat.}

\item{mixturefitmethod}{a string one of "circular", "axis-aligned", "general",
and "Bayesian" which determines the fitting method.("Bayesian" is not yet
supported)}

\item{param}{the number of components (in \code{list} form) for mixture
fitting and the concetnration parameter.}
}
\value{
returns an \code{icp.torus} object, containing all values
  to compute the conformity score.
}
\description{
\code{icp.torus.score} returns an icp.torus object, containing all values
  to compute the conformity score.
}
\examples{
\dontrun{
## mean vectors

Mu1 <- c(3, 0)
Mu2 <- c(2, 2)
Mu3 <- c(1, 4)

## covariance matrices

Sigma1 <- matrix(c(0.1, 0.05, 0.05, 0.2), 2, 2)
Sigma2 <- matrix(c(0.1, 0, 0, 0.01), 2, 2)
Sigma3 <- matrix(c(0.01, 0, 0, 0.1), 2, 2)

## 2-dimensional multivariate normal data wrapped with toroidal space
require(MASS)
data <- rbind(mvrnorm(n=70, Mu1, Sigma1),
              mvrnorm(n=50, Mu2, Sigma2),
              mvrnorm(n=50, Mu3, Sigma3))
data <- on.torus(data)

icp.torus <- icp.torus.score(data, method = "all",
                             mixturefitmethod = "general",
                             param = list(J = 4, concentration = 25))
}
}
\seealso{
\code{\link{EMsinvMmix}}, \code{\link[BAMBI]{dvmsinmix}}
}
