% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.torus.R
\name{kmeans.torus}
\alias{kmeans.torus}
\title{K-Means Clustering on Torus}
\usage{
kmeans.torus(
  data,
  centers = 10,
  method = c("extrinsic", "intrinsic"),
  iter.max = 100,
  nstart = 1
)
}
\arguments{
\item{data}{n x d matrix of toroidal data on \eqn{[0, 2\pi)^d}}

\item{centers}{either the number of clusters or a set of initial
cluster centers. If a number, a random set of row in x is
chosen as the initial centers.}

\item{method}{a string one of "intrinsic" or "extrinsic".
If "extrinsic", then extrinsic kmeans clustering method will be
implemented. ("intrinsic" is not yet supported)}

\item{iter.max}{the maximum number of iterations}

\item{nstart}{if \code{centers} is a number, how many random sets
should be chosen?}
}
\value{
returns a \code{kmeans} object, which contains
  input data, cluster centers on torus, membership,
  total sum of squares, within cluster sum of squares,
  between cluster centers sum of squares, and the size of
  each cluster.
}
\description{
\code{kmeans.torus} implements k-means clustering on toroidal
 space.
}
\details{
In Euclidean space, we know that the total sum of squares
  is equal to the summation of the within cluster sum of squares and
  the between cluster centers sum of squares. However, toroidal space
  does not satisfy the property; the equality does not hold. Thus,
  you need to be careful to use the sum of squares.
}
\examples{
\dontrun{
## mean vectors

Mu1 <- c(3, 0)
Mu2 <- c(2, 2)
Mu3 <- c(1, 4)

## covariance matrices

Sigma1 <- matrix(c(0.1, 0.05, 0.05, 0.2), 2, 2)
Sigma2 <- matrix(c(0.1, 0, 0, 0.01), 2, 2)
Sigma3 <- matrix(c(0.01, 0, 0, 0.1), 2, 2)

## 2-dimensional multivariate normal data wrapped with toroidal space
require(MASS)
data <- rbind(mvrnorm(n=70, Mu1, Sigma1),
              mvrnorm(n=50, Mu2, Sigma2),
              mvrnorm(n=50, Mu3, Sigma3))
data <- on.torus(data)

kmeans.torus(data, centers = 3, method = "extrinsic",
             iter.max = 100, nstart = 1)
}
}
\references{
'S. Jung, K. Park, and B. Kim (2020),
  "Clustering on the torus by conformal prediction", and
  Jaehyeok Shin, Alessandro Rinaldo and Larry Wasserman (2019),
  "Predictive Clustering"
}
\seealso{
\code{\link[stats]{kmeans}}, \code{\link[ClusTorus]{ang.minus}}
}
