% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mah.dist2.eval.R
\name{mah.dist2.eval}
\alias{mah.dist2.eval}
\title{Mahalanobis distance to ellipse}
\usage{
mah.dist2.eval(X, ellipse.param)
}
\arguments{
\item{X}{n x 2 toroidal data on \eqn{[0, 2\pi)^2}}

\item{ellipse.param}{list which is consisting of mean of each angular
coordinate, inverse of each covariance matrix, and constant term}
}
\value{
\code{mah.dist2.eval} returns  nrow(X) times ncol(parammat)
  (n x J) matrix.
}
\description{
\code{mah.dist2.eval} evaluates the mahalanobis distance from a toroidal
  point to each ellipses.
}
