% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.appr.param.R
\name{norm.appr.param}
\alias{norm.appr.param}
\title{Elliptical approximation of bivariate von Mises}
\usage{
norm.appr.param(parammat)
}
\arguments{
\item{parammat}{6 x J parameter data with the following components:
\code{parammat[1, ]} : the weights for each von Mises sine density
\code{parammat[n + 1, ]} : \eqn{\kappa_n} for each von Mises sine density for n = 1, 2, 3
\code{parammat[m + 4, ]} : \eqn{\mu_m} for each von Mises sine density for m = 1, 2}
}
\value{
returns approximated parameters for bivariate normal distribution with \code{list}:
  \code{list$Sigmainv[j]} : approximated covariance matrix for j-th von Mises
  \code{list$c[j]} : approximated \eqn{|2\pi\Sigma|^-1} for j-th von Mises
}
\description{
\code{norm.appr.param} approximates the bivariate von Mises sine densities
  to bivariate normal densities.
}
\seealso{
\code{\link[BAMBI]{dvmsin}}
}
