% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phat.eval.R
\name{phat.eval}
\alias{phat.eval}
\title{Product of weights and von Mises sine distribution functions}
\usage{
phat.eval(X, parammat)
}
\arguments{
\item{X}{n x 2 toroidal data on \eqn{[0, 2\pi)^2}}

\item{parammat}{6 x J parameter data with the following components:

  \code{parammat[1, ]} : the weights for each von Mises sine density

  \code{parammat[n + 1, ]} : \eqn{\kappa_n} for each von Mises
  sine density for n = 1, 2, 3

  \code{parammat[m + 4 , ]} : \eqn{\mu_m} for each von Mises
  sine density for m = 1, 2}
}
\value{
returns nrow(X) times ncol(parammat) (n x J) matrix
}
\description{
\code{phat.eval} evaluates \eqn{pi_j * f_j(x)}, where \eqn{pi_j} is a weight
  and \eqn{f_j(x)} is the j-th von Mises sine density which is determined by
  j-th parameters of the given parameter data.
}
\examples{
\dontrun{
## mean vectors

Mu1 <- c(3, 0)
Mu2 <- c(2, 2)
Mu3 <- c(1, 4)

## covariance matrices

Sigma1 <- matrix(c(0.1, 0.05, 0.05, 0.2), 2, 2)
Sigma2 <- matrix(c(0.1, 0, 0, 0.01), 2, 2)
Sigma3 <- matrix(c(0.01, 0, 0, 0.1), 2, 2)

## 2-dimensional multivariate normal data wrapped with toroidal space
require(MASS)
X <- rbind(mvrnorm(n=70, Mu1, Sigma1),
           mvrnorm(n=50, Mu2, Sigma2),
           mvrnorm(n=50, Mu3, Sigma3))

X <- on.torus(X)

parammat <- matrix(c(0.4, 0.3, 0.3,
                     20, 25, 25,
                     30, 25, 20,
                     1, 2, 3,
                     1, 2, 3,
                     0, 2, 4), nrow = 6, byrow = TRUE)

phat.eval(X, parammat)
}
}
\seealso{
\code{\link[BAMBI]{dvmsin}}
}
