% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.cluster.obj}
\alias{plot.cluster.obj}
\title{description}
\usage{
\method{plot}{cluster.obj}(
  x,
  method = c("outlier", "log.density", "posterior", "mahalanobis"),
  out = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{x} object}

\item{method}{A string. One of "outlier", "log.density", "posterior", "mahalanobis". Default is "outlier".}

\item{out}{An option for returning the ggplot object.}

\item{...}{additional parameter}
}
\description{
more details
}
