% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ellipsoids.R
\name{plot.ellipsoids}
\alias{plot.ellipsoids}
\title{Plot Ellipsoids on 2-dimensional Toroidal Space}
\usage{
\method{plot}{ellipsoids}(ellipse.param, t, coord = c(1, 2))
}
\arguments{
\item{t}{a numeric value which determines the size of ellipses.}

\item{coord}{a 2-vector for prespecifing the coordinates.
Default value is c(1, 2).}

\item{list}{which is consisting of mean of each angular
coordinate, inverse of each covariance matrix, and constant term}
}
\value{
a ggplot2-based 2-dimensional plot with ellipses
}
\description{
\code{plot.ellipsoids} generates a 2-dimensional plot for the given ellipsoids,
  with prespecified coordinates.
}
\examples{
\dontrun{
parammat <- matrix(c(0.4, 0.3, 0.3,
                     20, 25, 25,
                     30, 25, 20,
                     1, 2, 3,
                     1, 2, 3,
                     0, 2, 4), nrow = 6, byrow =TRUE)

ellipse.param <- norm.appr.param(parammat)

coord <- c(1, 3)
t <- 0.5

plot.ellipsoids(ellipse.param, t, coord)
}
}
\seealso{
\code{\link[ClusTorus]{icp.torus.score}}, \code{\link[ClusTorus]{cluster.assign.torus}}
}
