% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.icp.torus}
\alias{plot.icp.torus}
\title{description}
\usage{
\method{plot}{icp.torus}(x, data, level = 0.1, ellipse = TRUE, out = FALSE, ...)
}
\arguments{
\item{x}{\code{x} object}

\item{data}{\code{data}}

\item{level}{a scalar}

\item{ellipse}{A boolean index.}

\item{out}{A boolean index.}

\item{...}{additional parameter}
}
\description{
more details
}
