% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.icp.torus}
\alias{predict.icp.torus}
\title{description}
\usage{
\method{predict}{icp.torus}(object, newdata, ...)
}
\arguments{
\item{object}{\code{icp.torus} object}

\item{newdata}{\code{data}}

\item{...}{additional parameter}
}
\description{
more details
}
