% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.kmeans.torus}
\alias{predict.kmeans.torus}
\title{description}
\usage{
\method{predict}{kmeans.torus}(object, newdata, ...)
}
\arguments{
\item{newdata}{\code{data} to be input}

\item{...}{additional parameter}

\item{kmeans.torus}{\code{kmeans.torus} object}
}
\description{
more details
}
