% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protein_6a32}
\alias{protein_6a32}
\title{Protein data, for example}
\format{
A list with 11 attributes:
\describe{
  \item{psi}{angular variable, in arc-degree}
  \item{phi}{angular variable, in arc-degree}
  ...
}
}
\source{
\url{https://www.ebi.ac.uk/pdbe/entry/pdb/6A32}
}
\usage{
protein_6a32
}
\description{
A dataset containing 6 different angular data of almost 430 proteins.
}
\keyword{datasets}
