% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SARS_CoV_2}
\alias{SARS_CoV_2}
\title{6VXX: Structure of the SARS-CoV-2 spike glycoprotein(closed state)}
\format{
A list with 11 attributes:
\describe{
  \item{psi}{angular variable, in arc-degree}
  \item{phi}{angular variable, in arc-degree}
  ...
}
}
\source{
\url{http://www.rcsb.org/structure/6VXX}
}
\usage{
SARS_CoV_2
}
\description{
The torsion angle dataset of the SARS-CoV-2 spike glycopreotein.
}
\keyword{datasets}
