% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test.intersection.ellipse.R
\name{Test.intersection.ellipse}
\alias{Test.intersection.ellipse}
\title{Intersection of two ellipses on R^2}
\usage{
Test.intersection.ellipse(mean.1, M.1, mean.2, M.2)
}
\arguments{
\item{mean.1, mean.2}{d-dimensional vectors which indicate the
centers of each ellipse.}

\item{M.1, M.2}{d x d matrices which determine the shape of
each ellipse.}
}
\value{
If they intersect, then return \code{TRUE}. If not,
  then return \code{FALSE}.
}
\description{
\code{Test.intersection.ellipse} evaluates whether two ellipses
  on R^2 intersect.
}
\examples{
\dontrun{
mean.1 <- c(1, 0)
mean.2 <- c(0, 2)
M.1 <- diag(c(3, 4))
M.2 <- matrix(c(5, 2, 2, 1), nrow = 2, byrow = TRUE)

Test.intersection.ellipse(mean.1, M.1, mean.2, M.2)
}
}
\references{
We use the efficient algorithm given by
 David Eberly (2019), "Intersection of Ellipses",
 Geometric Tools, Redmond WA 98052
}
\seealso{
\code{\link[ClusTorus]{Test.intersection.ellipse.torus}}
}
