% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test.intersection.ellipse.torus.R
\name{Test.intersection.ellipse.torus}
\alias{Test.intersection.ellipse.torus}
\title{Intersection of two ellipses on torus}
\usage{
Test.intersection.ellipse.torus(ellipse.param, index, t)
}
\arguments{
\item{ellipse.param}{list which is consisting of mean of each angular
coordinate, inverse of each covariance matrix, and constant terms}

\item{index}{2-dimensional vector which indicates the ellipses that
we will check.}

\item{t}{a numeric value which determines the size of ellipses.}
}
\value{
If they intersect, then return \code{TRUE}. If not,
  then return \code{FALSE}.
}
\description{
\code{Test.intersection.ellipse.torus} evaluates whether two ellipses
  on torus intersect.
}
\examples{
\dontrun{
parammat <- matrix(c(0.4, 0.3, 0.3,
                     20, 25, 25,
                     30, 25, 20,
                     1, 2, 3,
                     1, 2, 3,
                     0, 2, 4), nrow = 6, byrow =TRUE)

elipse.param <- norm.appr.param(parammat)

index <- c(1, 3)
t <- 0.5

Test.intersection.ellipse.torus(ellipse.param, index, t)
}
}
\references{
S. Jung, K. Park, B. Kim (2020), "Clustering on the torus
  by conformal prediction"
}
\seealso{
\code{\link[ClusTorus]{Test.intersection.ellipse}}
}
