% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor.minus.R
\name{tor.minus}
\alias{tor.minus}
\title{Toroidal subtraction}
\usage{
tor.minus(data, mu)
}
\arguments{
\item{data}{n x 2 matrix of toroidal values}

\item{mu}{a 2-vector}
}
\description{
\code{tor.minus()} computes angular subtraction bewtween n x 2 toroidal data and
  a 2 dimensional vector.
}
\examples{
\dontrun{
data <- matrix(c(pi/3, pi/4, pi/2, -pi/2),
               ncol = 2, byrow = TRUE)
mu <- c(pi/2, -pi/2)

tor.minus(data, mu)
}
}
