% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toydata1}
\alias{toydata1}
\title{toydata1: Labelled Data for 5 Clusters}
\format{
This \code{data.frame} contains the following components:
\describe{
  \item{\code{phi}}{column for the first angle}
  \item{\code{psi}}{column for the second angle}
  \item{\code{label}}{column for the clustering membership}
}
}
\usage{
toydata1
}
\description{
Artificially generated data on the 2 dimensional torus
}
\details{
toydata1 is an artificial data generated from a mixture of 5 clusters,
  where three clusters are sampled from bivariate normal distributions
  and the other two are each sampled from the uniform distribution on a rectangle.
}
\references{
This simulation data is from S. Jung, K. Park, B. Kim (2021)
  "Clustering on the torus by conformal prediction". Annals of Applied Statistics
}
\keyword{datasets}
