% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toydata2}
\alias{toydata2}
\title{toydata2: Labelled Data for 3 Clusters}
\format{
This \code{data.frame} contains the following components:
\describe{
  \item{\code{phi}}{column for the first angle}
  \item{\code{psi}}{column for the second angle}
  \item{\code{label}}{column for the clustering membership}
}
}
\usage{
toydata2
}
\description{
Artificially generated data on the 2 dimensional torus
}
\details{
toydata2 is an artificial data generated from a mixture of 3 clusters,
  where the first cluster is sampled from a spherical normal distribution,
  the second cluster is from the uniform distribution on a large “L”-shaped region,
  and the third cluster of size 50 is sampled from the uniform distribution on the
  entire 2-dimensional torus.
}
\references{
This simulation data is from S. Jung, K. Park, B. Kim (2021)
  "Clustering on the torus by conformal prediction". Annals of Applied Statistics
}
\keyword{datasets}
