% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd.stat.amb.R
\name{wtd.stat.amb}
\alias{wtd.stat.amb}
\title{Weighted means in the ambient space and the cross matrix}
\usage{
wtd.stat.amb(data, w)
}
\arguments{
\item{data}{angular data whose elements are in \eqn{[0, 2\pi)}}

\item{w}{numeric vector whose each element is non-negative and
\code{sum(w) == 1}. Moreover, the length of \code{w} is the same with
\code{nrow(data)}.}
}
\value{
list which is consisting of \code{y1bar}, \code{y2bar}, \code{S12}.
}
\description{
See section A.1 of the article 'Clustering on the
  torus by conformal prediction', S. Jung, K. Park, and B. Kim (2020)
}
\details{
\code{wtd.stat.amb} returns the vector-valued weighted means in the
  ambient space and the cross matrix.
}
\references{
'S. Jung, K. Park, and B. Kim (2020),
  "Clustering on the torus by conformal prediction"
}
