% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ang.dist.R
\name{ang.dist}
\alias{ang.dist}
\title{Angular distances}
\usage{
ang.dist(x, y)
}
\arguments{
\item{x, y}{angular data on \eqn{[0, 2\pi)}}
}
\description{
\code{ang.dist()} computes angular distances.
}
\examples{
\dontrun{
x <- pi/3
y <- pi/4

ang.dist(x, y)
}
}
