% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ang.minus.R
\name{ang.minus}
\alias{ang.minus}
\title{Augular subtraction}
\usage{
ang.minus(x, y)
}
\arguments{
\item{x, y}{angular data on \eqn{[0, 2\pi)}}
}
\description{
\code{ang.minus()} computes angular subtraction.
}
\examples{
\dontrun{
x <- pi/2
y <- pi/3

ang.minus(x, y)
}
}
