% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.assign.torus.R
\name{cluster.assign.torus}
\alias{cluster.assign.torus}
\title{Clustering by connected components of ellipses}
\usage{
cluster.assign.torus(data, icp.torus, level)
}
\arguments{
\item{data}{n x 2 matrix of toroidal data on \eqn{[0, 2\pi)^2}.}

\item{icp.torus}{an object containing all values to compute the conformity
score, which will be constructed with \code{icp.torus.score}.}

\item{level}{either a scalar or a vector, or even \code{NULL}. Default value
is 0.1.}
}
\value{
clustering assignment for data, given icp.torus objects
}
\description{
\code{cluster.assign.torus} returns clustering assignment for data
  given \code{icp.torus} objects, which can be constructed with
  \code{icp.torus.score}.
}
\examples{
\dontrun{
## mean vectors

Mu1 <- c(3, 0)
Mu2 <- c(2, 2)
Mu3 <- c(1, 4)

## covariance matrices

Sigma1 <- matrix(c(0.1, 0.05, 0.05, 0.2), 2, 2)
Sigma2 <- matrix(c(0.1, 0, 0, 0.01), 2, 2)
Sigma3 <- matrix(c(0.01, 0, 0, 0.1), 2, 2)

## 2-dimensional multivariate normal data wrapped with toroidal space
require(MASS)
data <- rbind(mvrnorm(n=70, Mu1, Sigma1),
              mvrnorm(n=50, Mu2, Sigma2),
              mvrnorm(n=50, Mu3, Sigma3))
data <- on.torus(data)

icp.torus <- icp.torus.score(data, method = "all",
                             mixturefitmethod = "general",
                             param = list(J = 4, concentration = 25))
level <- c(0.1, 0.08)

cluster.assign.torus(data, icp.torus, level)
}
}
\references{
'S. Jung, K. Park, and B. Kim (2020),
  "Clustering on the torus by conformal prediction"
}
\seealso{
\code{\link[ClusTorus]{icp.torus.score}}
}
