% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehat.eval.R
\name{ehat.eval}
\alias{ehat.eval}
\title{Normal approximation to the log - weighted bivariate von Mises sine density}
\usage{
ehat.eval(X, ellipse.param)
}
\arguments{
\item{X}{n x 2 toroidal data on \eqn{[0, 2\pi)^2}}

\item{ellipse.param}{list which is consisting of mean of each angular
coordinate, inverse of each covariance matrix, and constant term}
}
\value{
nrow(X) times ncol(parammat) (n x J) matrix
}
\description{
\code{ehat.eval} evaluates the approximation of
  log - weighted bivariate von Mises for each given data and f
  or each given parameters.
}
\examples{
\dontrun{
## mean vectors

Mu1 <- c(3, 0)
Mu2 <- c(2, 2)
Mu3 <- c(1, 4)

## covariance matrices

Sigma1 <- matrix(c(0.1, 0.05, 0.05, 0.2), 2, 2)
Sigma2 <- matrix(c(0.1, 0, 0, 0.01), 2, 2)
Sigma3 <- matrix(c(0.01, 0, 0, 0.1), 2, 2)

## 2-dimensional multivariate normal data wrapped with toroidal space
require(MASS)
X <- rbind(mvrnorm(n=70, Mu1, Sigma1),
           mvrnorm(n=50, Mu2, Sigma2),
           mvrnorm(n=50, Mu3, Sigma3))

X <- on.torus(X)

parammat <- matrix(c(0.4, 0.3, 0.3,
                     20, 25, 25,
                     30, 25, 20,
                     1, 2, 3,
                     1, 2, 3,
                     0, 2, 4), nrow = 6, byrow =TRUE)

ellipse.param <- norm.appr.param(parammat)

ehat.eval(X, ellipse.param)
}
}
