% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.torus.R
\name{grid.torus}
\alias{grid.torus}
\title{Grid on torus}
\usage{
grid.torus(d = 2, grid.size = 100)
}
\arguments{
\item{d}{number for dimension. Default is 2.}

\item{grid.size}{number of grid for each axis. Default value is 100.}
}
\value{
returns (grid.size) x (grid.size) numeric matrix
  which indicates the grid points on torus.
}
\description{
\code{grid.torus} returns an equally-spaced grid on torus.
}
\examples{
grid.torus(d = 2, grid.size = 100)
}
