% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icp.torus.score.R
\name{icp.torus.score}
\alias{icp.torus.score}
\title{Inductive conformal prediction sets}
\usage{
icp.torus.score(
  data,
  split.id = NULL,
  method = c("all", "kde", "mixture"),
  mixturefitmethod = c("circular", "axis-aligned", "general", "Bayesian"),
  param = list(J = 4, concentration = 25)
)
}
\arguments{
\item{data}{n x 2 matrix of toroidal data on \eqn{[-\pi, \pi)^2}}

\item{split.id}{a n-vector consisting of values 1 (estimation) and 2(evaluation)}

\item{method}{a string one of "all", "kde" and "mixture" which determines the
phat.}

\item{mixturefitmethod}{a string one of "circular", "axis-aligned", "general",
and "Bayesian" which determines the fitting method.("Bayesian" is not yet
supported)}

\item{param}{the number of components (in \code{list} form) for mixture
fitting and the concetnration parameter.}
}
\value{
returns an icp.torus object, containing all values
  to compute the conformity score
}
\description{
\code{icp.torus.score} returns an icp.torus object, containing all values
  to compute the conformity score.
}
\examples{
\dontrun{
data <- matrix(c(-pi/3, -pi/3, pi/2, pi/4),
               nrow = 2, byrow = TRUE)

icp.torus.score(data, method = "kde", mixturefitmethod = "circular")
}
}
\seealso{
\code{\link{EMsinvMmix}}, \code{\link[BAMBI]{dvmsinmix}}
}
