% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.kspheres.R
\name{kmeans.kspheres}
\alias{kmeans.kspheres}
\title{K-Means Clustering to K-Spheres Clustering on Torus}
\usage{
kmeans.kspheres(
  data,
  centers = 10,
  type = c("homogeneous-circular", "heterogeneous-circular", "general"),
  parammat = EMsinvMmix.init(data, centers),
  THRESHOLD = 1e-10,
  maxiter = 200,
  verbose = TRUE
)
}
\arguments{
\item{data}{data n x d matrix of toroidal data on \eqn{[0, 2\pi)^d}}

\item{centers}{either the number of clusters or a set of initial
cluster centers. If a number, a random set of row in x is
chosen as the initial centers.}

\item{type}{character which must be "homogeneous-circular",
"heterogeneous-circular", or "general".
 If "homogeneous-circular", the radii of k-spheres are identical.
 If "heterogeneous-cricular", the radii of k-spheres may be different.
 If, "general", clustering with k-ellipses. The parameters to construct
 the ellipses are optimized with generalized Lloyd algorithm, which is
 modified for toroidal space. To see the detail, see the references.}
}
\value{
returns a \code{sphere.param} object,
  containing all values which determines the shape and
  location of spheres.
}
\description{
\code{kmeans.kspheres} prepares the parameters for conformity scores
  which are derived by k-means clustering on torus.
}
\examples{
\dontrun{
## mean vectors

Mu1 <- c(3, 0)
Mu2 <- c(2, 2)
Mu3 <- c(1, 4)

## covariance matrices

Sigma1 <- matrix(c(0.1, 0.05, 0.05, 0.2), 2, 2)
Sigma2 <- matrix(c(0.1, 0, 0, 0.01), 2, 2)
Sigma3 <- matrix(c(0.01, 0, 0, 0.1), 2, 2)

## 2-dimensional multivariate normal data wrapped with toroidal space
require(MASS)
data <- rbind(mvrnorm(n=70, Mu1, Sigma1),
              mvrnorm(n=50, Mu2, Sigma2),
              mvrnorm(n=50, Mu3, Sigma3))
data <- on.torus(data)

kmeans.kspheres(data, centers = 3, type = "various")
}
}
\references{
S. Jung, K. Park, and B. Kim (2020),
  "Clustering on the torus by conformal prediction", and
  Jaehyeok Shin, Alessandro Rinaldo and Larry Wasserman (2019),
  "Predictive Clustering"
}
\seealso{
\code{\link[ClusTorus]{ehat.eval}},
  \code{\link[ClusTorus]{kmeans.torus}}
}
