% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mah.dist2.eval.R
\name{mah.dist2.eval}
\alias{mah.dist2.eval}
\title{Mahalanobis distance to ellipse}
\usage{
mah.dist2.eval(X, ellipse.param)
}
\arguments{
\item{X}{n x 2 toroidal data on \eqn{[0, 2\pi)^2}}

\item{ellipse.param}{list which is consisting of mean of each angular
coordinate, inverse of each covariance matrix, and constant term}
}
\value{
\code{mah.dist2.eval} returns nrow(X) times ncol(parammat)
  (n x J) matrix whose entries implies the Mahalanobis distance
  between i-th data point and the j-th ellipse.
}
\description{
\code{mah.dist2.eval} evaluates the mahalanobis distance from a toroidal
  point to each ellipses.
}
\examples{
\dontrun{
## mean vectors

Mu1 <- c(3, 0)
Mu2 <- c(2, 2)
Mu3 <- c(1, 4)

## covariance matrices

Sigma1 <- matrix(c(0.1, 0.05, 0.05, 0.2), 2, 2)
Sigma2 <- matrix(c(0.1, 0, 0, 0.01), 2, 2)
Sigma3 <- matrix(c(0.01, 0, 0, 0.1), 2, 2)

## 2-dimensional multivariate normal data wrapped with toroidal space
require(MASS)
X <- rbind(mvrnorm(n=70, Mu1, Sigma1),
           mvrnorm(n=50, Mu2, Sigma2),
           mvrnorm(n=50, Mu3, Sigma3))

X <- on.torus(X)

parammat <- matrix(c(0.4, 0.3, 0.3,
                     20, 25, 25,
                     30, 25, 20,
                     1, 2, 3,
                     1, 2, 3,
                     0, 2, 4), nrow = 6, byrow =TRUE)

elipse.param <- norm.appr.param(parammat)

mah.dist2.eval(X, ellipse.param)
}
}
