% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phat.eval.R
\name{phat.eval}
\alias{phat.eval}
\title{Product of weights and von Mises sine distribution functions}
\usage{
phat.eval(X, parammat)
}
\arguments{
\item{X}{n x 2 toroidal data on \eqn{[0, 2\pi)^2}}

\item{parammat}{6 x J parameter data with the following components:
\code{parammat[1, ]} : the weights for each von Mises sine density
\code{parammat[n + 1, ]} : \eqn{\kappa_n} for each von Mises sine density for n = 1, 2, 3
\code{parammat[m + 4 , ]} : \eqn{\mu_m} for each von Mises sine density for m = 1, 2}
}
\value{
returns nrow(X) times ncol(parammat) (n x J) matrix
}
\description{
\code{phat.eval} evaluates \eqn{pi_j * f_j(x)}, where \eqn{pi_j} is a weight
  and \eqn{f_j(x)} is the j-th von Mises sine density which is determined by
  j-th parameters of the given parameter data.
}
\seealso{
\code{\link[BAMBI]{dvmsin}}
}
