% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SARS_CoV_2}
\alias{SARS_CoV_2}
\title{6VXX: Structure of the SARS-CoV-2 spike glycoprotein(closed state)}
\format{
This data.frame contains the following columns:
\describe{
  \item{\code{phi}}{main chain torsion angle for atoms C,N,CA,C.}
  \item{\code{psi}}{main chain torsion angle for atoms N,CA,C,N.}
  \item{\code{type}}{type of protein chains}
}
}
\source{
This data can be downloaded in
  \url{https://www.rcsb.org/structure/6VXX}, or with using R package
  \code{bio3d}.
}
\usage{
SARS_CoV_2
}
\description{
The torsion angle dataset of the SARS-CoV-2 spike glycopreotein.
}
\references{
Walls, A.C., et al. (2020), "Structure of the SARS-CoV-2
  spike glycoprotein (closed state)" Cell 181: 281, DOI:10.2210/pdb6vxx/pdb.
  Retrived from \url{https://www.wwpdb.org/pdb?id=pdb_00006vxx}
}
\seealso{
Description of the angluar information is from the 'value'
  part of \code{torsion.pdb} in the package \code{bio3d}.
}
\keyword{datasets}
