% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor.minus.R
\name{tor.minus}
\alias{tor.minus}
\title{Toroidal subtraction}
\usage{
tor.minus(data, mu)
}
\arguments{
\item{data}{n x d matrix of toroidal data}

\item{mu}{a d-dimensinal vector}
}
\value{
angular subtraction bewtween n x d toroidal data and
  a d dimensional vector.
}
\description{
\code{tor.minus} computes angular subtraction bewtween n x d toroidal data and
  a d dimensional vector.
}
\examples{
\dontrun{
data <- matrix(c(pi/3, pi/4, pi/2, pi/2),
               ncol = 2, byrow = TRUE)
mu <- c(pi/2, pi/2)

tor.minus(data, mu)
}
}
\seealso{
\code{\link{ang.minus}}
}
